+++ dotnet/Tests/Eyes.Images.E2ETests/ImagesGenericTestsBase.cs
@@ -1,6 +1,8 @@
 using System;
-using System.Threading;
 using Applitools;
+using Applitools.Metadata;
+using Applitools.Tests.Utils;
+using Applitools.Utils;
 using NUnit.Framework;
 
 namespace Eyes.Images.E2ETests
@@ -8,7 +10,7 @@ namespace Eyes.Images.E2ETests
     [Parallelizable(ParallelScope.Children)]
     public abstract class ImagesGenericTestsBase
     {
-        private readonly ThreadLocal<Applitools.Images.Eyes> eyes_ = new();
+        private readonly System.Threading.ThreadLocal<Applitools.Images.Eyes> eyes_ = new();
         private static string TEST_SUITE_NAME = "Eyes Image SDK";
         protected static BatchInfo Batch = new BatchInfo(TEST_SUITE_NAME);
 
@@ -42,6 +44,23 @@ namespace Eyes.Images.E2ETests
         protected string GetApplicationName()
         {
             return "Applitools Eyes SDK";
+        }
+        
+        protected SessionResults GetTestInfo(TestResults results)
+        {
+            SessionResults sessionResults = null;
+            try
+            {
+                sessionResults = TestUtils.GetSessionResults(Eyes.ApiKey, results);
+            }
+            catch (Exception e)
+            {
+                //CommonUtils.LogExceptionStackTrace(logger_, Stage.TestFramework, StageType.TestResults, e);
+                Assert.Fail("Exception appeared while getting session results");
+            }
+
+            ArgumentGuard.NotNull(sessionResults, nameof(sessionResults));
+            return sessionResults;
         }
     }
 }