+++ dotnet/Eyes.Images.DotNet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Bug Fixes
+
+* Fixed `BatchInfo.AddProperty` ([Trello 3280](https://trello.com/c/Z3RpuEns))
+
 ## [3.42](https://github.com/applitools/eyes.sdk.javascript1/compare/dotnet/images@3.41...dotnet/images@3.42) (2023-08-27)
 
 ### Features